backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
altair==4.2.0
annoy==1.17.1
anyio==3.6.1
arrow==1.2.3
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
binaryornot==0.4.4
black==22.10.0
blobfile==2.0.1
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cohere==2.8.0
cookiecutter==1.7.3
cryptography==39.0.2
dataclasses-json==0.5.7
datasets==2.6.1
dill==0.3.5.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.85.0
filelock==3.8.0
frozenlist==1.3.3
fsspec==2022.11.0
greenlet==1.1.3
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.9.0
isort==5.12.0
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
jsonschema==4.17.0
langchain==0.0.74
libcst==0.4.9
llvmlite==0.39.1
loguru==0.6.0
lxml==4.9.2
Mako==1.2.3
manage-fastapi==1.1.1
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.2
multiprocess==0.70.13
mypy-extensions==0.4.3
numba==0.56.4
numpy==1.21.6
openai==0.23.1
openpyxl==3.0.10
orjson==3.8.5
packaging==21.3
pandas==1.3.5
pandas-stubs==1.2.0.62
passlib==1.7.4
pathspec==0.10.1
pinecone-client==2.0.13
platformdirs==2.5.3
poyo==0.5.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
pyarrow==10.0.0
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.2
pynndescent==0.5.8
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==6.1.2
pytz==2022.4
PyYAML==6.0
questionary==1.10.0
redis==4.4.2
regex==2022.10.31
requests==2.28.1
responses==0.18.0
rfc3986==1.5.0
rsa==4.9
ruff==0.0.243
scikit-learn==1.1.3
scipy==1.9.3
sentry-sdk==1.13.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sse-starlette==1.2.1
starlette==0.20.4
text-unidecode==1.3
threadpoolctl==3.1.0
tiktoken==0.2.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
typer==0.4.2
typing-inspect==0.8.0
typing_extensions==4.3.0
umap-learn==0.5.3
unimport==0.14.1
urllib3==1.26.12
uvicorn==0.18.3
wcwidth==0.2.5
xxhash==3.1.0
yarl==1.8.1
zipp==3.8.1
